Fixed designators and clarified footprint of D2.
I did change the C2 and R4 designators since they were mixed up. 
I also did modify the footprint of D2 to make it more obvious how to orient it since usually pin 1 is marked on the physical package but for that LED it's pin 3.
Also updated the files to KiCad 7.